fixed sudden death after exit bug
Fixed bug where the user would pause and "finish" the game during a
sudden death round, and then return, and it would still be sudden death.
This bug isn't entirely accessible from the iSSB windows version, as
there is no pause screen, but I still felt the need to fix it.